Resolved conflicts in KV/Hybrid prefix cache Match signatures by composing
both new params: lora_id (this branch, per-adapter namespacing) and intent
(main, distinguishes PrefixReuse from StateRecovery for retracted-request
recovery). Both call sites in forward.cpp (scheduleDecodeFromRetracted and
the post-allocation re-match) now pass request->LoraId() together with
MatchIntent::StateRecovery so retracted LoRA requests recover from their
own adapter namespace.

Also merged ForwardContext: kept the new last_index_offsets field from
main alongside the lora_manager field on this branch.

+++ python/tokenspeed/runtime/execution/runtime_states.py
@@ -62,9 +62,7 @@ class RuntimeStates:
     def update_valid_cache_length(
         self, req_pool_indices: torch.Tensor, increment_lengths: torch.Tensor
     ) -> None:
-        self.valid_cache_lengths.index_add_(
-            0, req_pool_indices.long(), increment_lengths
-        )
+        self.valid_cache_lengths.index_add_(0, req_pool_indices, increment_lengths)
 
     def reset_states(
         self,
